+++ src/app/api/solana/action/register/route.ts
@@ -3,10 +3,10 @@ import { compose } from '@masknet/shared-base';
 import { createSuccessResponseJSON } from '@/helpers/createSuccessResponseJSON.js';
 import { fetchJSON } from '@/helpers/fetchJSON.js';
 import { withRequestErrorHandler } from '@/helpers/withRequestErrorHandler.js';
-import type { ActionsRegistryConfig } from '@/providers/types/Blink.js';
+import type { ActionsRegistryResponse } from '@/providers/types/Blink.js';
 
 export const GET = compose(withRequestErrorHandler(), async () => {
-    const config = await fetchJSON<ActionsRegistryConfig>('https://actions-registry.dialectapi.to/all', {
+    const config = await fetchJSON<ActionsRegistryResponse>('https://actions-registry.dialectapi.to/all', {
         next: {
             revalidate: 60,
         },
