+++ src/store/useComposeStore.ts
@@ -19,9 +19,9 @@ import { createPoll } from '@/helpers/polls.js';
 import { BlinkLoader } from '@/providers/blink/Loader.js';
 import { FrameLoader } from '@/providers/frame/Loader.js';
 import { OpenGraphLoader } from '@/providers/og/Loader.js';
-import type { Action } from '@/providers/types/Blink.js';
 import type { CompositePoll } from '@/providers/types/Poll.js';
 import type { Channel, Post } from '@/providers/types/SocialMedia.js';
+import type { Action } from '@/types/blink.js';
 import { type ComposeType, type MediaObject } from '@/types/compose.js';
 import type { Frame } from '@/types/frame.js';
 import type { OpenGraph } from '@/types/og.js';
